refac: Order를 app store receipt에 맞게 수정
