--- src/main/java/atwoz/atwoz/order/domain/OrderStatus.java
@@ -1,9 +1,0 @@
-package atwoz.atwoz.order.domain;
-
-public enum OrderStatus {
-    REQUESTED,
-    PAID,
-    COMPLETED,
-    REFUNDED,
-    CANCELED
-}
