+++ src/main/java/atwoz/atwoz/payment/application/exception/OrderAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package atwoz.atwoz.payment.application.exception;
+
+public class OrderAlreadyExistsException extends RuntimeException {
+    public OrderAlreadyExistsException() {
+        super("이미 존재하는 주문입니다.");
+    }
+}
