+++ src/main/java/atwoz/atwoz/payment/domain/OrderStatus.java
@@ -1,0 +1,6 @@
+package atwoz.atwoz.payment.domain;
+
+public enum OrderStatus {
+    PAID,
+    REFUNDED
+}
